+++ matching-service/package.json
@@ -2,10 +2,10 @@
   "name": "matching-service",
   "version": "1.0.0",
   "description": "",
-  "main": "index.ts",
+  "main": "src/index.ts",
   "scripts": {
-    "dev": "nodemon index.ts",
-    "start": "ts-node index.ts",
+    "dev": "nodemon src/index.ts",
+    "start": "ts-node src/index.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
